requirements.txt
Django==1.7.4
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Unipath==1.1
Werkzeug==0.9.6
bottle==0.12.8
chardet==2.3.0
decorator==3.4.0
dj-database-url==0.3.0
dj-static==0.0.6
django-csvimport==2.4
django-dajaxice==0.7
django-debug-toolbar==1.2.2
django-social-auth==0.7.28
django-toolbelt==0.0.1
gunicorn==19.2.1
httplib2==0.9
itsdangerous==0.24
matplotlib==1.4.2
networkx==1.9.1
numpy==1.9.1
oauth2==1.5.211
psycopg2==2.6
pyparsing==2.0.3
python-dateutil==2.3
python-instagram==1.1.3
python-openid==2.2.5
pytz==2014.10
redis==2.10.3
simplejson==3.6.5
six==1.9.0
sqlparse==0.1.14
static3==0.5.1
virtualenv==12.0.7
-r requirements_heroku.txt